refactor: 아이디 찾기 메서드 위치 수정 #42
