+++ src/main/java/com/beauty4u/backend/user/command/domain/repository/UserRepository.java
@@ -6,5 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<UserInfo, String> {
+
     Optional<UserInfo> findByUserCode(String userCode);
+
+    Optional<UserInfo> findByUserNameAndPhoneAndEmail(String userName, String phone, String email);
 }
