+++ src/main/java/com/beauty4u/backend/user/query/mapper/UserQueryMapper.java
@@ -12,9 +12,4 @@ public interface UserQueryMapper {
     List<UserListResDTO> findUserList(
             @Param("offset") Long offset,
             @Param("count") Long count);
-
-    String findUserCode(
-            @Param("name") String name,
-            @Param("phone") String phone,
-            @Param("email") String email);
 }
